Split WSQ benchmark into header file and client code
